Добавляет возможность выбора времени
